Converted spkwriter app
