+++ isis/src/base/apps/spkwriter/spkwriter.h
@@ -1,0 +1,10 @@
+#ifndef spkwriter_h
+#define spkwriter_h
+
+#include "UserInterface.h"
+#include "Pvl.h"
+namespace Isis {
+  void spkwriter(UserInterface &ui, Pvl *log);
+}
+
+#endif
